+++ src/router/router.tsx
@@ -1,16 +1,16 @@
 import App from '../App';
-import MyAccount from '../pages/MyAccount/MyAccount';
 import ErrorPage from '../pages/ErrorPage/ErrorPage';
 import About from '../pages/About/About';
 import EighthNotesMenu from '../pages/EighthNotesMenu/EighthNotesMenu';
 import TripletNotesMenu from '../pages/TripletNotesMenu/TripletNotesMenu';
 import RandomNotesMenu from '../pages/RandomNotesMenu/RandomNotesMenu';
-import { createBrowserRouter } from 'react-router-dom';
-import { routes } from './routes';
 import MenuLayout from '../pages/MenuLayout/MenuLayout';
 import RedirectToEighths from '../pages/RedirectToEighths/RedirectToEighths';
-import Login from '../pages/MyAccount/Login/Login';
-import UserAccount from '../pages/MyAccount/UserAccount/UserAccount';
+import Login from '../pages/Login/Login';
+import Account from '../pages/Account/Account';
+
+import { createBrowserRouter } from 'react-router-dom';
+import { routes } from './routes';
 
 export const router = createBrowserRouter([
   {
@@ -31,15 +31,7 @@ export const router = createBrowserRouter([
         ],
       },
       { path: routes.login, element: <Login /> },
-      { path: routes.account, element: <UserAccount /> },
-      // {
-      //   path: routes.account,
-      //   element: <MyAccount />,
-      //   children: [
-      //     { path: 'login', element: <Login /> },
-      //     { path: ':userId', element: <UserAccount /> },
-      //   ],
-      // },
+      { path: routes.account, element: <Account /> },
       { path: routes.about, element: <About /> },
     ],
   },
